switch to mb

+++ src/vivarium_cluster_tools/psimulate/results/cli_options.py
@@ -12,7 +12,7 @@ import click
 from vivarium_cluster_tools.cli_tools import MINUTES_OR_NONE
 
 DEFAULT_BATCH_SIZE = 200
-DEFAULT_OUTPUT_FILE_SIZE_MB = 500
+DEFAULT_OUTPUT_FILE_SIZE_MB = 500.0
 
 with_batch_size = click.option(
     "--batch-size",
@@ -24,11 +24,11 @@ with_batch_size = click.option(
 with_output_file_size = click.option(
     "--output-file-size",
     "output_file_size",
-    type=int,
+    type=float,
     default=DEFAULT_OUTPUT_FILE_SIZE_MB,
     show_default=True,
     help="Maximum file size in MB per result file (within each type of result). When exceeded, a new file is started.",
-    callback=lambda ctx, param, value: value * 1024 * 1024,  # Convert MB to bytes
+    callback=lambda ctx, param, value: int(value * 1024 * 1024),  # Convert MB to bytes
 )
 with_no_batch = click.option(
     "--no-batch",
